Transfer more methods to MaskImageWidget (needed by RoiStackPlugin)
